nlp/requirements.txt
fasttext
spacy
tqdm
unidecode
pyspellchecker
gcld3